requirements.txt
wheel
requests
certifi
confluent-kafka[avro,json,protobuf]>=1.4.2
